extract load balancer fwded IP detection
This functionality is not actually core to sseserver, so we shouldn't
enable it by default.  Extract to a public http.Handler, which we hook
in by default in our Serve() convenience method, but it will not
interfere with the chain build by someone making their own middleware
chain unless they specifically request it.

It iss possible to totally remove this from this package in the future,
and possibly just make it it's own package, but for now let's
preserve it in Server for simplicity.