introduce optimized routing data structure v1
Short on documentation and tests, and most likely contains grievous
implementation errors in the corner cases since we aren't testing for
them yet, but don't want to invest any time there until we're sure we
can actually make a worthwhile difference here.

This is a case of "replace one line of code with hundreds of lines of
code" so it needs to really really be worth it!