refactor: remove (*Server).Serve and middleware
In order to simplify this module, remove anything custom that was
originally added here for convenience. Using Server as a standard
http.Handler is the preferred way to do things, and lets people use
their own middleware for logging, RealIP management, etc.

Really that was unnecessary "batteries included" best left to web
frameworks (which all have their own middleware included for standard
stuff like that, so rely on the application to do the work.