refactor: handle server config via options pattern
This changes the signature of NewServer() to potentially return an error.

Currently, there are no cases where an error would be returned, and this
will always be nil, but changing the signature in our v2 refactor gives
us flexibility in the future if there are failable situations during
Server initialization.