refactor(hub): switch from chans to mutex
performance impact appears negligible based on benchmarks in hub_test,
however this allows for a somewhat less complex internal implementation
than the previous "actor model" channel model.

As a bonus, this removes the need for "ensures delivery is finished"
calls to Shutdown() as a temporary hack in the hub tests.